Split logic from spancat recipe across several files
This makes it so that the functions used in the recipe can be
imported without need for prodigy to be available, e.g. in tests.
