+++ scripts/recipes/spancat.py
@@ -1,279 +1,20 @@
-import re
 from functools import partial
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    Iterable,
-    Iterator,
-    Optional,
-    Union,
-    List,
-    Container,
-)
-from collections import defaultdict
+from typing import Any, Callable, Container, Dict, Iterable, Iterator, Optional, Union
 
 import prodigy
 import spacy
-import srsly
 from nltk.util import ngrams
 from prodigy.components.loaders import get_stream
-from prodigy.components.preprocess import add_tokens
 from prodigy.types import RelationsTask
 from prodigy.util import set_hashes, split_string
+from spacy.language import Language
 from spacy.pipeline.spancat import Suggester
 from spacy.tokens import Doc, Span
-from spacy.matcher import PhraseMatcher
-from spacy.language import Language
 from thinc.api import Ops, get_current_ops
 from thinc.types import Ragged
 
-
-if not Span.has_extension("atomic"):
-    Span.set_extension("atomic", default=False)
-
-if not Span.has_extension("possible_entity"):
-    Span.set_extension("possible_entity", default=True)
-
-SPAN_LABELS = ["SEM", "GRAF", "PHON", "META", "PER", "WORK"]
-"""Labels for types of content in Jingdian Shiwen annotations."""
-
-REL_LABELS = ["SRC", "MOD"]
-"""Labels for types of relations between spans in Jingdian Shiwen annotations."""
-
-MODIFIER = "上下又並同一或亦後末"
-MODIFIER2 = "注皆章及文篇卦"
-MARKER = "作云也同音無"
-WORK = "書本文言子注經卦詩"
-
-SPAN_PATTERN_MAP = {
-    "PHON": [
-        re.compile(rf"^[{MODIFIER}]*?音?(.)(.)之\1|\2$"),
-        re.compile(rf"^[{MODIFIER}]*?音?..反$"),
-        re.compile(rf"^[{MODIFIER}]*?音?如字$"),
-        re.compile(rf"^[{MODIFIER}]*?音[^{MARKER}]$"),
-    ],
-    "SEM": [
-        re.compile(rf"^[{MODIFIER}]*?[^{MARKER}]+也$"),
-    ],
-    "GRAF": [
-        re.compile(rf"^[{MODIFIER}]*?[作無][^{MARKER}{MODIFIER}]+$"),
-    ],
-    "META": [
-        re.compile(rf"^[{MODIFIER}{MODIFIER2}]+[^{MARKER}]*同$"),
-        re.compile(r"^出注$"),
-        re.compile(r"^絶句$"),
-        re.compile(r"^字非$"),
-    ],
-    "MARKER": [
-        re.compile(rf"^[{MODIFIER}]*?[{MARKER}]$"),
-    ],
-}
-
-ENT_PATTERN_MAP = {
-    "WORK": [
-        re.compile(rf"^.*?[{WORK}]$"),
-    ],
-    "PER": [
-        re.compile(rf"^[^{WORK}{MARKER}]{{1,3}}$"),
-    ],
-}
-
-SPAN_PATTERNS = [
-    pattern for patterns in SPAN_PATTERN_MAP.values() for pattern in patterns
-]
-
-ENT_PATTERNS = [
-    pattern for patterns in ENT_PATTERN_MAP.values() for pattern in patterns
-]
-
-PHON_PATTERN = re.compile(rf"([{MODIFIER}]*?音?(?:(?:..反)|音.|如字))")
-XYZY_PATTERN = re.compile(r"音?(.)(.)之\1|\2")
-SPLIT_AFTER = re.compile(r"(.+?[也同]+)")
-SPLIT_AROUND = re.compile(rf"([{MODIFIER}]*?[云])")
-SPLIT_BEFORE = re.compile(rf"([{MODIFIER}]*?[作無][^{MARKER}]+)")
-SPLIT_BEFORE_2 = re.compile(rf"([^{MARKER}{MODIFIER}{MODIFIER2}]+)(同)")
-
-# pre-defined entity patterns
-patterns = defaultdict(list)
-for pattern in srsly.read_jsonl("assets/jdsw_ner_patterns.jsonl"):
-    patterns[pattern["label"]].append(pattern["pattern"])
-
-# set up a PhraseMatcher for pre-defined entities
-nlp = spacy.blank("zh")
-MATCHER = PhraseMatcher(nlp.vocab, None)
-for label, _patterns in patterns.items():
-    MATCHER.add(label, [nlp.make_doc(pattern) for pattern in _patterns])
-
-
-def check_span(span: Span) -> Span:
-    """Check whether a span contains entities and/or is atomic, and set the flags accordingly."""
-    # spans of length 1 can no longer be split
-    if len(span) == 1:
-        span._.atomic = True
-
-    # check if the span matches a known span pattern
-    # if so, label it and mark it as atomic
-    # also mark it as not containing any entities
-    for label, patterns in SPAN_PATTERN_MAP.items():
-        if any(pattern.fullmatch(span.text) for pattern in patterns):
-            span.label_ = label
-            span._.atomic = True
-            span._.possible_entity = False
-            return span
-
-    # check if the span matches a known entity pattern
-    # if so, label it and mark it as atomic
-    matches = MATCHER(span.as_doc())
-    if len(matches) == 1 and (matches[0][2] - matches[0][1] == len(span)):
-        span.label_ = nlp.vocab.strings[matches[0][0]]
-        span._.atomic = True
-        span._.possible_entity = True
-        return span
-
-    return span
-
-
-def split_spans(
-    spans: Iterable[Span],
-    split_fn: Callable[[str], List[str]],
-) -> List[Span]:
-    """
-    Use a provided function to split each non-atomic span, yielding a flat list of spans.
-    """
-    output_spans = []
-    _split_fn = split_fn
-
-    def _split_span(span: Span) -> List[Span]:
-        output_spans = []
-        subspans = list(filter(bool, _split_fn(span.text)))
-
-        # must be a non-destructive split
-        assert "".join(subspans) == span.text
-
-        subspan_start = 0
-        for subspan in subspans:
-            output_spans.append(
-                span.doc.char_span(
-                    span.start + subspan_start,
-                    span.start + subspan_start + len(subspan),
-                )
-            )
-            subspan_start += len(subspan)
-
-        return output_spans
-
-    for span in spans:
-        if span._.atomic:
-            output_spans.append(span)
-        else:
-            for subspan in _split_span(span):
-                output_spans.append(check_span(subspan))
-
-    return output_spans
-
-
-def split_at_indices(text: str, indices: List[int]) -> List[str]:
-    """Split a string at the given indices."""
-    return [
-        part
-        for part in (text[i:j] for i, j in zip([0] + indices, indices + [None]))  # type: ignore
-        if part
-    ]
-
-
-def split_backref_noncapture(text: str, pattern: re.Pattern) -> List[str]:
-    """Split a string using a regex with backreferences, but don't capture the backreferences."""
-    indices = []
-    for match in pattern.finditer(text):
-        indices.append(match.start())
-        indices.append(match.end())
-    return split_at_indices(text, sorted(list(set(indices))))
-
-
-def split_phrase_matcher(doc: Doc, matcher: PhraseMatcher) -> List[str]:
-    """Split a string using a provided PhraseMatcher instance."""
-    indices = []
-    for match in matcher(doc):
-        indices.append(match[1])
-        indices.append(match[2])
-    return split_at_indices(doc.text, sorted(list(set(indices))))
-
-
-def doc_spans_jdsw(doc: Doc) -> Iterable[Span]:
-    """Split a Jingdian Shiwen annotation into (non-overlapping) labeled spans."""
-    # start with the entire doc as a single span
-    spans = [doc.char_span(0, len(doc.text))]
-
-    # pass 1: phonetic patterns
-    spans = split_spans(spans, PHON_PATTERN.split)
-    spans = split_spans(spans, lambda s: split_backref_noncapture(s, XYZY_PATTERN))
-
-    # pass 2: known entities
-    spans = split_spans(spans, lambda s: split_phrase_matcher(nlp.make_doc(s), MATCHER))
-
-    # pass 2: span-final characters
-    spans = split_spans(spans, SPLIT_AFTER.split)
-
-    # pass 3: span-separator characters
-    spans = split_spans(spans, SPLIT_AROUND.split)
-
-    # pass 4: span-initial characters
-    spans = split_spans(spans, SPLIT_BEFORE.split)
-    spans = split_spans(spans, SPLIT_BEFORE_2.split)  # ent + 同
-
-    # pass 5: catchall entity detection
-    for span in spans:
-        for label, patterns in ENT_PATTERN_MAP.items():
-            if any(pattern.fullmatch(span.text) for pattern in patterns):
-                if not span.label_:  # don't overwrite existing labels
-                    span.label_ = label
-                    span._.atomic = True
-                    span._.possible_entity = True
-
-    return spans
-
-
-def span_rels_jdsw(spans: List[Span]) -> Iterable[Dict[str, Any]]:
-    """Extract relations from Jingdian Shiwen spans."""
-    rels: List[Dict[str, Any]] = []
-    subtrees: List[List[Span]] = [[]]
-
-    # first pass: add relations between meta spans and what they modify
-    for i, span in enumerate(spans):
-        if i > 0 and span.label_ == "META" and spans[i - 1].label_ in SPAN_LABELS:
-            rels.append(
-                {
-                    "head": span.end - 1,
-                    "child": spans[i - 1].end - 1,
-                    "label": "MOD",
-                }
-            )
-
-    # second pass: add relations between spans and their textual sources (entities)
-    while len(spans):
-        span = spans.pop(0)
-        if span.label_ not in ["PER", "WORK"]:
-            subtrees[-1].append(span)
-        else:
-            subtrees.append([span])
-    for subtree in subtrees:
-        if len(subtree) < 2:
-            continue
-        head = subtree[0]
-        if head.label_ not in ["PER", "WORK"]:
-            continue
-        for span in subtree[1:]:
-            if span.label_ in ["SEM", "GRAF", "PHON"]:
-                rels.append(
-                    {
-                        "head": head.end - 1,
-                        "child": span.end - 1,
-                        "label": "SRC",
-                    }
-                )
-
-    return rels
+from scripts.lib.components import doc_spans_jdsw, span_rels_jdsw
+from scripts.lib.patterns import REL_LABELS, SPAN_LABELS
 
 
 def span_ngram_suggester(
